commands: Move replacement selection support
Include and discard both use replacement-selection behavior when a command
expands or validates line selections around mixed replacement runs.

That behavior lived inside include even though discard imported one of the
helpers directly. That made include act as both a command module and shared
command-selection support.

This commit addresses that by moving replacement-selection helpers into
commands.selection.replacement_selection and updating include and discard
to import the support module.

The next commit will move tests that exercise replacement run analysis to
the support module API.
